+++ src/version2/prototype/processor/Clip.java
@@ -9,7 +9,6 @@ import org.gdal.gdal.Band;
 import org.gdal.gdal.Dataset;
 import org.gdal.gdal.Transformer;
 import org.gdal.gdal.gdal;
-import org.gdal.gdalconst.gdalconst;
 import org.gdal.gdalconst.gdalconstConstants;
 import org.gdal.ogr.DataSource;
 import org.gdal.ogr.Layer;
@@ -80,7 +79,7 @@ public class Clip
             }
         }
         // remove the input folder
-        if(deleteInputDirectory) {
+        if(deleteInputDirectory && inputFolder.exists()) {
             FileUtils.deleteDirectory(inputFolder);
         }
     }
