+++ src/version2/prototype/processor/Convert.java
@@ -54,7 +54,7 @@ public abstract class Convert {
         convertFiles();
 
         // remove the input folder
-        if(deleteInputDirectory) {
+        if(deleteInputDirectory && inputFolder.exists()) {
             FileUtils.deleteDirectory(inputFolder);
         }
