+++ src/version2/prototype/processor/Mask.java
@@ -84,7 +84,7 @@ public class Mask {
         }
 
         // remove the input folder
-        if(deleteInputDirectory) {
+        if(deleteInputDirectory && inputFolder.exists()) {
             FileUtils.deleteDirectory(inputFolder);
         }
     }
