+++ src/version2/prototype/processor/Reproject.java
@@ -76,7 +76,7 @@ public abstract class Reproject {
         }
 
         // remove the input folder
-        if(deleteInputDirectory) {
+        if(deleteInputDirectory && inputFolder.exists()) {
             FileUtils.deleteDirectory(inputFolder);
         }
     }
